Add TimeModel class
- Refactor TimeParser

- Add TimeModel class

- Move methods not related to parsing date
into TimeModel

- Create test file for the TimeModel class

- Move test methods that were moved into TimeModel
into the test file for TimeModel
